+++ include/lwipopts.h
@@ -70,4 +70,43 @@
 
 #define LWIP_PROVIDE_ERRNO              1
 
+
+// meine Versuche
+#define LWIP_TCPIP_CORE_LOCKING         1
+
+
+
+
+// Debugging (was sonst?)
+#define LWIP_DEBUG
+
+#define API_LIB_DEBUG                   LWIP_DBG_ON
+#define API_MSG_DEBUG                   LWIP_DBG_ON
+#define AUTOIP_DEBUG                    LWIP_DBG_ON
+#define DHCP_DEBUG                      LWIP_DBG_ON
+#define DNS_DEBUG                       LWIP_DBG_ON
+#define ETHARP_DEBUG                    LWIP_DBG_ON
+#define ICMP_DEBUG                      LWIP_DBG_ON
+#define IGMP_DEBUG                      LWIP_DBG_ON
+#define INET_DEBUG                      LWIP_DBG_ON
+#define IP_DEBUG                        LWIP_DBG_ON
+#define NETIF_DEBUG                     LWIP_DBG_ON
+#define PBUF_DEBUG                      LWIP_DBG_ON
+#define RAW_DEBUG                       LWIP_DBG_ON
+#define SLIP_DEBUG                      LWIP_DBG_ON
+#define SOCKETS_DEBUG                   LWIP_DBG_ON
+#define SYS_DEBUG                       LWIP_DBG_ON
+#define TCP_DEBUG                       LWIP_DBG_ON
+#define TCP_INPUT_DEBUG                 LWIP_DBG_ON
+#define TCP_FR_DEBUG                    LWIP_DBG_ON
+#define TCP_RTO_DEBUG                   LWIP_DBG_ON
+#define TCP_CWND_DEBUG                  LWIP_DBG_ON
+#define TCP_WND_DEBUG                   LWIP_DBG_ON
+#define TCP_RST_DEBUG                   LWIP_DBG_ON
+#define TCP_OUTPUT_DEBUG                LWIP_DBG_ON
+#define TCP_QLEN_DEBUG                  LWIP_DBG_ON
+#define TCPIP_DEBUG                     LWIP_DBG_ON
+#define TIMERS_DEBUG                    LWIP_DBG_OFF
+#define UDP_DEBUG                       LWIP_DBG_ON
+
 #endif /* __LWIPOPTS_H__ */
